Getting going documentation (#78)
Add an overview of the software dependencies, and a quick start guide
for running an LFRic app. The latter is based on the existing wiki page
(which can be removed once the documentation here is published).

The software dependencies page is currently the starting point for
core1.2. In the medium term, I envisage the software dependencies page
will be enhanced by release notes identifying changes to the software
dependencies and reasons for the changes.